requirements/base.txt
python-dateutil==2.7.2
Django>=1.11,<2.1
django-swingtime==0.9.0

